pyproject.toml
[tool.poetry]
name = "advertisements_back"
version = "0.1.0"
description = "Веб-приложение для продажи б/у вещей"
authors = ["David Golovanov"]
license = "Private"
readme = "README.md"
packages = [{include = "*", from="src"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "5.2.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
django-filter = "25.1"
python-dotenv = "^1.0.1"
pillow = "11.2.1"
psycopg2 = "^2.9.10"
setuptools = "80.8.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
django-stubs = "^5.0.2"
djangorestframework-stubs = "^3.15.0"

[tool.poetry.group.deploy.dependencies]
gunicorn = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
